requirements.txt
flask
pymongo
python-dotenv
pandas
google-generativeai
pyjwt
bcrypt